% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{breakthrough}
\alias{breakthrough}
\title{breakthrough}
\usage{
breakthrough(...)

breakthrough(...)
}
\arguments{
\item{...}{arguments to forward to the vroom::vroom function.}
}
\description{
Breakthrough by State
Breakthrough data scraped from multiple states daily. Contains information on vaccinated vs unvaccinated hospitalizations and COVID cases.
More information available at the \href{https://epi.quartzsoftware.com/datasets/breakthrough}{Quartz Software Infection Disease Data Repository}

Breakthrough by State
Breakthrough data scraped from multiple states daily. Contains information on vaccinated vs unvaccinated hospitalizations and COVID cases.
More information available at the \href{https://epi.quartzsoftware.com/datasets/breakthrough}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated

This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated
}
\examples{
data <- breakthrough()

data <- breakthrough()

}
