% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{caserates.by.age}
\alias{caserates.by.age}
\title{caserates.by.age}
\usage{
caserates.by.age(save = F)
}
\description{
COVID19 Caserates by Age Group
Weekly case rates of COVID 19 by age group as reported by the CDC
More information available at the \href{https://epi.quartzsoftware.com/datasets/caserates-by-age}{Quartz Software Infection Disease Data Repository}
}
\examples{
data <- caserates.by.age()

}
