% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{covid.by.county}
\alias{covid.by.county}
\title{covid.by.county}
\usage{
covid.by.county(save = F)
}
\description{
COVID by County
US COVID 19 community levels by county as reported by the CDC
More information available at the \href{https://epi.quartzsoftware.com/datasets/covid-by-county}{Quartz Software Infection Disease Data Repository}
}
\examples{
data <- covid.by.county()

}
