% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{covid_by_county}
\alias{covid_by_county}
\title{covid_by_county}
\usage{
covid_by_county(...)
}
\arguments{
\item{...}{arguments to forward to the vroom::vroom function.}
}
\description{
COVID by County
US COVID 19 community levels by county as reported by the CDC
More information available at the \href{https://epi.quartzsoftware.com/datasets/covid-by-county}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated
}
\examples{
data <- covid_by_county()

}
