% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{cprd}
\alias{cprd}
\title{cprd}
\usage{
cprd(use_memory = T, use_disk = T, bust_cache = F, ...)

cprd(use_memory = T, use_disk = T, bust_cache = F, ...)
}
\arguments{
\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}
}
\description{
CPRD
All protocols scraped from the cprd website
More information available at the \href{https://epi.quartzsoftware.com/datasets/cprd}{Quartz Software Infection Disease Data Repository}

CPRD
All protocols scraped from the cprd website
More information available at the \href{https://epi.quartzsoftware.com/datasets/cprd}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

This function uses the cache_download function to download the data from the repository.
}
\examples{
data <- cprd()

data <- cprd()

}
