% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{deathcounts.by.age}
\alias{deathcounts.by.age}
\title{deathcounts.by.age}
\usage{
deathcounts.by.age()
}
\description{
COVID19 Death Counts by Age Group
This datasets is pulled daily from the CDC's data tracker for easier access. It provides weekly death counts secondary to COVID 19 stratified by age group.
More information available at the \href{https://epi.quartzsoftware.com/datasets/deathcounts-by-age}{Quartz Software Infection Disease Data Repository}
}
\examples{
data <- deathcounts.by.age()

}
