% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{deathcounts_by_age}
\alias{deathcounts_by_age}
\title{deathcounts_by_age}
\usage{
deathcounts_by_age(use_memory = T, use_disk = T, bust_cache = F, ...)
}
\arguments{
\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}
}
\description{
COVID19 Death Counts by Age Group
This datasets is pulled daily from the CDC's data tracker for easier access. It provides weekly death counts secondary to COVID 19 stratified by age group.
More information available at the \href{https://epi.quartzsoftware.com/datasets/deathcounts-by-age}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.
}
\examples{
data <- deathcounts_by_age()

}
