% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{deathrates.by.age}
\alias{deathrates.by.age}
\title{deathrates.by.age}
\usage{
deathrates.by.age()
}
\description{
COVID19 Death Rates by Age Group
Weekly death rates of COVID 19 by age group as reported by the CDC
More information available at the \href{https://epi.quartzsoftware.com/datasets/deathrates-by-age}{Quartz Software Infection Disease Data Repository}
}
\examples{
data <- deathrates.by.age()

}
