% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{flu}
\alias{flu}
\title{flu}
\usage{
flu(use_memory = T, use_disk = T, bust_cache = F, ...)

flu(use_memory = T, use_disk = T, bust_cache = F, ...)
}
\arguments{
\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}
}
\description{
Flu Vaccination Rates
Flu vaccination rates stratified by race, county, and season.
More information available at the \href{https://epi.quartzsoftware.com/datasets/flu}{Quartz Software Infection Disease Data Repository}

Flu Vaccination Rates
Flu vaccination rates stratified by race, county, and season.
More information available at the \href{https://epi.quartzsoftware.com/datasets/flu}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

This function uses the cache_download function to download the data from the repository.
}
\examples{
data <- flu()

data <- flu()

}
