% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{flu.testing}
\alias{flu.testing}
\title{flu.testing}
\usage{
flu.testing(...)
}
\arguments{
\item{...}{arguments to forward to the vroom::vroom function.}
}
\description{
Flu Testing
Age Group Distribution of Influenza Positive Specimens Reported by Public Health Laboratories
More information available at the \href{https://epi.quartzsoftware.com/datasets/flu-testing}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated
}
\examples{
data <- flu.testing()

}
