% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{flunet}
\alias{flunet}
\title{flunet}
\usage{
flunet(use_memory = T, use_disk = T, bust_cache = F, ...)

flunet(use_memory = T, use_disk = T, bust_cache = F, ...)
}
\arguments{
\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}

\item{...}{arguments to forward to the cache_download function.}
}
\description{
WHO Influenza
All of the WHO's influenza data describing case counts by influenza subtype by country and year.
More information available at the \href{https://epi.quartzsoftware.com/datasets/flunet}{Quartz Software Infection Disease Data Repository}

WHO Influenza
All of the WHO's influenza data describing case counts by influenza subtype by country and year.
More information available at the \href{https://epi.quartzsoftware.com/datasets/flunet}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

This function uses the cache_download function to download the data from the repository.
}
\examples{
data <- flunet()

data <- flunet()

}
