% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{hospitalizations.by.age}
\alias{hospitalizations.by.age}
\title{hospitalizations.by.age}
\usage{
hospitalizations.by.age()
}
\description{
COVID19 Hospitalizations by Age Group
Weekly hospitalizations of COVID 19 by age group as reported by the CDC
More information available at https://epi.quartzsoftware.com/datasets/hospitalizations-by-age
}
\examples{
data <- hospitalizations.by.age()

}
