% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{influenza.burden}
\alias{influenza.burden}
\title{influenza.burden}
\usage{
influenza.burden(use_memory = T, use_disk = T, bust_cache = F, ...)
}
\arguments{
\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}

\item{...}{arguments to forward to the cache_download function.}
}
\description{
Influenza Burden
Seasonal data since 2010 with multiple measures to compare influenza burden by season.
More information available at the \href{https://epi.quartzsoftware.com/datasets/influenza-burden}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.
}
\examples{
data <- influenza.burden()

}
