% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{influenza.prevention}
\alias{influenza.prevention}
\title{influenza.prevention}
\usage{
influenza.prevention(save = F)
}
\description{
Influenza Prevention
Influenza prevention data from CDC's Influenza Prevention and Surveillance Program.
More information available at the \href{https://epi.quartzsoftware.com/datasets/influenza-prevention}{Quartz Software Infection Disease Data Repository}
}
\examples{
data <- influenza.prevention()

}
