% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{influenza.vaccine.effectiveness}
\alias{influenza.vaccine.effectiveness}
\title{influenza.vaccine.effectiveness}
\usage{
influenza.vaccine.effectiveness(
  use_memory = T,
  use_disk = T,
  bust_cache = F,
  ...
)
}
\arguments{
\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}

\item{...}{arguments to forward to the cache_download function.}
}
\description{
Influenza Vaccine Effectiveness
Seasonal data since 2010 with vaccine effectiveness by age group.
More information available at the \href{https://epi.quartzsoftware.com/datasets/influenza-vaccine-effectiveness}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.
}
\examples{
data <- influenza.vaccine.effectiveness()

}
