% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{influenza_burden}
\alias{influenza_burden}
\title{influenza_burden}
\usage{
influenza_burden(...)
}
\arguments{
\item{...}{arguments to forward to the vroom::vroom function.}
}
\description{
Influenza Burden
Seasonal data since 2010 with multiple measures to compare influenza burden by season.
More information available at the \href{https://epi.quartzsoftware.com/datasets/influenza-burden}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated
}
\examples{
data <- influenza_burden()

}
