% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{influenza_prevention}
\alias{influenza_prevention}
\title{influenza_prevention}
\usage{
influenza_prevention(...)
}
\arguments{
\item{...}{arguments to forward to the vroom::vroom function.}
}
\description{
Influenza Prevention
Influenza prevention data from CDC's Influenza Prevention and Surveillance Program.
More information available at the \href{https://epi.quartzsoftware.com/datasets/influenza-prevention}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated
}
\examples{
data <- influenza_prevention()

}
