% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{misc}
\alias{misc}
\title{misc}
\usage{
misc(...)

misc(...)
}
\arguments{
\item{...}{arguments to forward to the vroom::vroom function.}
}
\description{
MISC Data
Daily data on MISC cases vs COVID cases in the US.
More information available at the \href{https://epi.quartzsoftware.com/datasets/misc}{Quartz Software Infection Disease Data Repository}

MISC Data
Daily data on MISC cases vs COVID cases in the US.
More information available at the \href{https://epi.quartzsoftware.com/datasets/misc}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated

This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated
}
\examples{
data <- misc()

data <- misc()

}
