% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{nervss}
\alias{nervss}
\title{nervss}
\usage{
nervss(use_memory = T, use_disk = T, bust_cache = F, ...)

nervss(use_memory = T, use_disk = T, bust_cache = F, ...)
}
\arguments{
\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}

\item{...}{arguments to forward to the cache_download function.}
}
\description{
NERVSS
NERVSS respiratory pathogen testing information
More information available at the \href{https://epi.quartzsoftware.com/datasets/nervss}{Quartz Software Infection Disease Data Repository}

NERVSS
NERVSS respiratory pathogen testing information
More information available at the \href{https://epi.quartzsoftware.com/datasets/nervss}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

This function uses the cache_download function to download the data from the repository.
}
\examples{
data <- nervss()

data <- nervss()

}
