% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{outpatient.viral.surveillance}
\alias{outpatient.viral.surveillance}
\title{outpatient.viral.surveillance}
\usage{
outpatient.viral.surveillance()
}
\description{
Outpatient Viral Surveillance
National, Regional, and State Level Outpatient Illness and Viral Surveillance data pulled monthly from the cdc.
More information available at the \href{https://epi.quartzsoftware.com/datasets/outpatient-viral-surveillance}{Quartz Software Infection Disease Data Repository}
}
\examples{
data <- outpatient.viral.surveillance()

}
