% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{syndromic.trends}
\alias{syndromic.trends}
\title{syndromic.trends}
\usage{
syndromic.trends(save = F)
}
\description{
Syndromic Trends
Trends of different infectious pathogens in the US over time.
More information available at the \href{https://epi.quartzsoftware.com/datasets/syndromic-trends}{Quartz Software Infection Disease Data Repository}
}
\examples{
data <- syndromic.trends()

}
