% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{test.burden}
\alias{test.burden}
\title{test.burden}
\usage{
test.burden()
}
\description{
Test Burden
Test burden data from the CDC.
More information available at the \href{https://epi.quartzsoftware.com/datasets/test-burden}{Quartz Software Infection Disease Data Repository}
}
\examples{
data <- test.burden()

}
