% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{test_burden}
\alias{test_burden}
\title{test_burden}
\usage{
test_burden(use_memory = T, use_disk = T, bust_cache = F, ...)
}
\arguments{
\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}

\item{...}{arguments to forward to the cache_download function.}
}
\description{
Test Burden
Test burden data from the CDC.
More information available at the \href{https://epi.quartzsoftware.com/datasets/test-burden}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.
}
\examples{
data <- test_burden()

}
