% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{who.influenza}
\alias{who.influenza}
\title{who.influenza}
\usage{
who.influenza()
}
\description{
WHO Influenza
All of the WHO's influenza data describing case counts by influenza subtype by country and year.
More information available at https://epi.quartzsoftware.com/datasets/who-influenza
}
\examples{
data <- who.influenza()

}
