% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{who_influenza}
\alias{who_influenza}
\title{who_influenza}
\usage{
who_influenza(use_memory = T, use_disk = T, bust_cache = F, ...)
}
\arguments{
\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}
}
\description{
WHO Influenza
All of the WHO's influenza data describing case counts by influenza subtype by country and year.
More information available at the \href{https://epi.quartzsoftware.com/datasets/who-influenza}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.
}
\examples{
data <- who_influenza()

}
