% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{breakthrough}
\alias{breakthrough}
\title{breakthrough}
\usage{
breakthrough()
}
\description{
Breakthrough by State
Breakthrough data scraped from multiple states daily. Contains information on vaccinated vs unvaccinated hospitalizations and COVID cases.
More information available at https://epi.quartzsoftware.com/datasets/breakthrough
}
\examples{
data <- breakthrough()

}
