% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_download.R
\name{cache_download}
\alias{cache_download}
\title{Cache Download}
\usage{
cache_download(
  url,
  ...,
  use_memory = TRUE,
  use_disk = TRUE,
  bust_cache = FALSE
)
}
\arguments{
\item{url}{The URL to download from.}

\item{...}{arguments to forward to the vroom::vroom function.}

\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}
}
\description{
By default, all downloads are cached both on the disk and in your computer memory.
This means that the second time you download a file during the same session it will be reloaded from your ram.
If you stop a session and restart it, the file will be loaded from your disk.
If you want to reload the data from its original source, use the bust_cache function.
}
\details{
Additionally, options can be used to determine the default caching behavior of tidyiddr.
}
\examples{
cache_download("https://s3.amazonaws.com/quartzdata/datasets/caserates-by-age.csv")

options("tidyiddr_use_cache" = TRUE)
cache_download("https://s3.amazonaws.com/quartzdata/datasets/caserates-by-age.csv")
}
