% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{covariants}
\alias{covariants}
\title{covariants}
\usage{
covariants(...)

covariants(...)
}
\arguments{
\item{...}{arguments to forward to the vroom::vroom function.}
}
\description{
COVID19 Covariants by State
CoVariants provides an overview of SARS-CoV-2 variants and mutations that are of interest. Here, you can find out what mutations define a variant, what impact they might have (with links to papers and resources), and where variants are found
More information available at the \href{https://epi.quartzsoftware.com/datasets/covariants}{Quartz Software Infection Disease Data Repository}

COVID19 Covariants by State
CoVariants provides an overview of SARS-CoV-2 variants and mutations that are of interest. Here, you can find out what mutations define a variant, what impact they might have (with links to papers and resources), and where variants are found
More information available at the \href{https://epi.quartzsoftware.com/datasets/covariants}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated

This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated
}
\examples{
data <- covariants()

data <- covariants()

}
