% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{covariants}
\alias{covariants}
\title{covariants}
\usage{
covariants()
}
\description{
COVID19 Covariants by State
CoVariants provides an overview of SARS-CoV-2 variants and mutations that are of interest. Here, you can find out what mutations define a variant, what impact they might have (with links to papers and resources), and where variants are found
More information available at https://epi.quartzsoftware.com/datasets/covariants
}
\examples{
data <- covariants()

}
