% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{cprd}
\alias{cprd}
\title{cprd}
\usage{
cprd()
}
\description{
CPRD
All protocols scraped from the cprd website
More information available at https://epi.quartzsoftware.com/datasets/cprd
}
\examples{
data <- cprd()

}
