% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{deathrates_by_age}
\alias{deathrates_by_age}
\title{deathrates_by_age}
\usage{
deathrates_by_age(use_memory = T, use_disk = T, bust_cache = F, ...)
}
\arguments{
\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}
}
\description{
COVID19 Death Rates by Age Group
Weekly death rates of COVID 19 by age group as reported by the CDC
More information available at the \href{https://epi.quartzsoftware.com/datasets/deathrates-by-age}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.
}
\examples{
data <- deathrates_by_age()

}
