% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{flu}
\alias{flu}
\title{flu}
\usage{
flu(...)

flu(...)
}
\arguments{
\item{...}{arguments to forward to the vroom::vroom function.}
}
\description{
Flu Vaccination Rates
Flu vaccination rates stratified by race, county, and season.
More information available at the \href{https://epi.quartzsoftware.com/datasets/flu}{Quartz Software Infection Disease Data Repository}

Flu Vaccination Rates
Flu vaccination rates stratified by race, county, and season.
More information available at the \href{https://epi.quartzsoftware.com/datasets/flu}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated

This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated
}
\examples{
data <- flu()

data <- flu()

}
