% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{flu}
\alias{flu}
\title{flu}
\usage{
flu()
}
\description{
Flu Vaccination Rates
Flu vaccination rates stratified by race, county, and season.
More information available at the \href{https://epi.quartzsoftware.com/datasets/flu}{Quartz Software Infection Disease Data Repository}
}
\examples{
data <- flu()

}
