% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{flu.testing}
\alias{flu.testing}
\title{flu.testing}
\usage{
flu.testing(save = F)
}
\description{
Flu Testing
Age Group Distribution of Influenza Positive Specimens Reported by Public Health Laboratories
More information available at the \href{https://epi.quartzsoftware.com/datasets/flu-testing}{Quartz Software Infection Disease Data Repository}
}
\examples{
data <- flu.testing()

}
