% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{flunet}
\alias{flunet}
\title{flunet}
\usage{
flunet(...)

flunet(...)
}
\arguments{
\item{...}{arguments to forward to the vroom::vroom function.}
}
\description{
WHO Influenza
All of the WHO's influenza data describing case counts by influenza subtype by country and year.
More information available at the \href{https://epi.quartzsoftware.com/datasets/flunet}{Quartz Software Infection Disease Data Repository}

WHO Influenza
All of the WHO's influenza data describing case counts by influenza subtype by country and year.
More information available at the \href{https://epi.quartzsoftware.com/datasets/flunet}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated

This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated
}
\examples{
data <- flunet()

data <- flunet()

}
