% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{hospitalizations_by_age}
\alias{hospitalizations_by_age}
\title{hospitalizations_by_age}
\usage{
hospitalizations_by_age(use_memory = T, use_disk = T, bust_cache = F, ...)
}
\arguments{
\item{use_memory}{Whether to use memory caching.}

\item{use_disk}{Whether to use disk caching.}

\item{bust_cache}{Whether to bust (refresh) the cache.}

\item{...}{arguments to forward to the cache_download function.}
}
\description{
COVID19 Hospitalizations by Age Group
Weekly hospitalizations of COVID 19 by age group as reported by the CDC
More information available at the \href{https://epi.quartzsoftware.com/datasets/hospitalizations-by-age}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.
}
\examples{
data <- hospitalizations_by_age()

}
