% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{influenza.burden}
\alias{influenza.burden}
\title{influenza.burden}
\usage{
influenza.burden()
}
\description{
Influenza Burden
Seasonal data since 2010 with multiple measures to compare influenza burden by season.
More information available at https://epi.quartzsoftware.com/datasets/influenza-burden
}
\examples{
data <- influenza.burden()

}
