% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{influenza.vaccine.effectiveness}
\alias{influenza.vaccine.effectiveness}
\title{influenza.vaccine.effectiveness}
\usage{
influenza.vaccine.effectiveness()
}
\description{
Influenza Vaccine Effectiveness
Seasonal data since 2010 with vaccine effectiveness by age group.
More information available at https://epi.quartzsoftware.com/datasets/influenza-vaccine-effectiveness
}
\examples{
data <- influenza.vaccine.effectiveness()

}
