% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{influenza.vaccine.effectiveness}
\alias{influenza.vaccine.effectiveness}
\title{influenza.vaccine.effectiveness}
\usage{
influenza.vaccine.effectiveness(...)
}
\arguments{
\item{...}{arguments to forward to the vroom::vroom function.}
}
\description{
Influenza Vaccine Effectiveness
Seasonal data since 2010 with vaccine effectiveness by age group.
More information available at the \href{https://epi.quartzsoftware.com/datasets/influenza-vaccine-effectiveness}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated
}
\examples{
data <- influenza.vaccine.effectiveness()

}
