% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{misc}
\alias{misc}
\title{misc}
\usage{
misc()
}
\description{
MISC Data
Daily data on MISC cases vs COVID cases in the US.
More information available at the \href{https://epi.quartzsoftware.com/datasets/misc}{Quartz Software Infection Disease Data Repository}
}
\examples{
data <- misc()

}
