% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated.R
\name{outpatient.viral.surveillance}
\alias{outpatient.viral.surveillance}
\title{outpatient.viral.surveillance}
\usage{
outpatient.viral.surveillance(...)
}
\arguments{
\item{...}{arguments to forward to the vroom::vroom function.}
}
\description{
Outpatient Viral Surveillance
National, Regional, and State Level Outpatient Illness and Viral Surveillance data pulled monthly from the cdc.
More information available at the \href{https://epi.quartzsoftware.com/datasets/outpatient-viral-surveillance}{Quartz Software Infection Disease Data Repository}
}
\details{
This function uses the cache_download function to download the data from the repository.

Use use_memory, use_disk, and bust_cache have been deprecated
}
\examples{
data <- outpatient.viral.surveillance()

}
